+++ chitra/serve/__init__.py
@@ -1,1 +1,2 @@
-from chitra.serve.model_server import API, create_api, get_available_api_types
+from chitra.serve.api import API, create_api
+from chitra.serve.model_server import get_available_api_types
